Refactoring domain class.
