+++ meetapp/meetapp/pom.xml
@@ -8,14 +8,14 @@
 		<version>2.7.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>br.com.spring.meetapp</groupId>
-	<artifactId>meetapp</artifactId>
+	<groupId>br.com.celtic.banking</groupId>
+	<artifactId>celtic-banking</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>war</packaging>
-	<name>meetapp</name>
-	<description>meetapp</description>
+	<name>celtic-banking</name>
+	<description>Mock Bank App </description>
 	<properties>
-		<java.version>17</java.version>
+		<java.version>11</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
