+++ meetapp/meetapp/src/main/java/br/com/celtic/bank/domain/Client.java
@@ -1,0 +1,7 @@
+package br.com.celtic.bank.domain;
+
+import lombok.Data;
+
+@Data
+public class Client {
+}
