+++ meetapp/meetapp/src/main/java/br/com/celtic/bank/request/ClientRequest.java
@@ -1,0 +1,11 @@
+package br.com.celtic.bank.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ClientRequest {
+    private String agency;
+    private String account;
+}
