+++ meetapp/meetapp/src/main/java/br/com/celtic/bank/start/CelticBankingApplication.java
@@ -1,4 +1,4 @@
-package br.com.spring.meetapp.start;
+package br.com.celtic.bank.start;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,9 +7,9 @@ import org.springframework.context.annotation.ComponentScan;
 import static java.util.List.of;
 
 @SpringBootApplication
-@ComponentScan(basePackages = "br.com.spring.meetapp")
-public class MeetApp {
+@ComponentScan(basePackages = {"br.com.celtic.bank"})
+public class CelticBankingApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(MeetApp.class, args);
+		SpringApplication.run(CelticBankingApplication.class, args);
 	}
 }
