+++ meetapp/meetapp/src/main/java/br/com/celtic/bank/start/ServletInitializer.java
@@ -1,4 +1,4 @@
-package br.com.spring.meetapp.start;
+package br.com.celtic.bank.start;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
@@ -7,6 +7,6 @@ public class ServletInitializer extends SpringBootServletInitializer {
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(MeetApp.class);
+		return application.sources(CelticBankingApplication.class);
 	}
 }
