--- meetapp/meetapp/src/main/java/br/com/spring/pendency/domain/Pendency.java
@@ -1,10 +1,0 @@
-package br.com.spring.meetapp.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Agent {
-    private String name;
-}
