+++ meetapp/meetapp/src/main/resources/application.yml
@@ -1,4 +1,1 @@
-server:
-  port: 9999
-  servlet:
-    context-path: '/api'
+
