+++ meetapp/meetapp/src/test/java/br/com/celtic/bank/CelticBankingApplicationApplicationTests.java
@@ -1,10 +1,10 @@
-package br.com.spring.meetapp;
+package br.com.celtic.bank;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class MeetAppApplicationTests {
+class CelticBankingApplicationApplicationTests {
 
 	@Test
 	void contextLoads() {
